Fixed bug: Addition Min was stored as int in database (now decimal). Also updated fields to decimal to prevent potential rounding problems.

+++ Synologen.Core/Domain/Model/Orders/EyeParameter.cs
@@ -2,7 +2,7 @@ namespace Spinit.Wpc.Synologen.Core.Domain.Model.Orders
 {
 	public class EyeParameter
 	{
-		public virtual float? Left { get; set; }
-		public virtual float? Right { get; set; }
+		public virtual decimal? Left { get; set; }
+		public virtual decimal? Right { get; set; }
 	}
 }
