+++ Synologen.Presentation.Intranet/Models/Orders/CreateOrderModel.cs
@@ -54,18 +54,18 @@ namespace Spinit.Wpc.Synologen.Presentation.Intranet.Models.Orders
         public int SelectedArticleId { get; set; }
         public int SelectedShippingOption { get; set; }
 
-        public float SelectedLeftPower { get; set; }
-        public float SelectedLeftBaseCurve { get; set; }
-        public float SelectedLeftDiameter { get; set; }
-        public float SelectedLeftCylinder { get; set; }
-        public float SelectedLeftAxis { get; set; }
-        public float SelectedLeftAddition { get; set; }
-        public float SelectedRightPower { get; set; }
-        public float SelectedRightBaseCurve { get; set; }
-        public float SelectedRightDiameter { get; set; }
-        public float SelectedRightCylinder { get; set; }
-        public float SelectedRightAxis { get; set; }
-        public float SelectedRightAddition { get; set; }
+        public decimal SelectedLeftPower { get; set; }
+        public decimal SelectedLeftBaseCurve { get; set; }
+        public decimal SelectedLeftDiameter { get; set; }
+        public decimal SelectedLeftCylinder { get; set; }
+        public decimal SelectedLeftAxis { get; set; }
+        public decimal SelectedLeftAddition { get; set; }
+        public decimal SelectedRightPower { get; set; }
+        public decimal SelectedRightBaseCurve { get; set; }
+        public decimal SelectedRightDiameter { get; set; }
+        public decimal SelectedRightCylinder { get; set; }
+        public decimal SelectedRightAxis { get; set; }
+        public decimal SelectedRightAddition { get; set; }
 
 		public IEnumerable<ListItem> Categories { get; set; }
     	public IEnumerable<ListItem> Suppliers { get; set; }
